+++ tools/confix/migrations.go
@@ -19,15 +19,13 @@ const (
 // MigrationMap defines a mapping from a version to a transformation plan.
 type MigrationMap map[string]func(from *tomledit.Document, to string) transform.Plan
 
-var (
-	Migrations = MigrationMap{
-		"v0.45": NoPlan, // Confix supports only the current supported SDK version. So we do not support v0.44 -> v0.45.
-		"v0.46": PlanBuilder,
-		"v0.47": PlanBuilder,
-		"v0.48": PlanBuilder,
-		// "v0.xx.x": PlanBuilder, // add specific migration in case of configuration changes in minor versions
-	}
-)
+var Migrations = MigrationMap{
+	"v0.45": NoPlan, // Confix supports only the current supported SDK version. So we do not support v0.44 -> v0.45.
+	"v0.46": PlanBuilder,
+	"v0.47": PlanBuilder,
+	"v0.48": PlanBuilder,
+	// "v0.xx.x": PlanBuilder, // add specific migration in case of configuration changes in minor versions
+}
 
 // PlanBuilder is a function that returns a transformation plan for a given diff between two files.
 func PlanBuilder(from *tomledit.Document, to string) transform.Plan {
@@ -58,7 +56,8 @@ func PlanBuilder(from *tomledit.Document, to string) transform.Plan {
 								Block: parser.Comments{
 									"###############################################################################",
 									fmt.Sprintf("###							%s Configuration							###", strings.Title(kv.Key)),
-									"###############################################################################"},
+									"###############################################################################",
+								},
 								Name: keys,
 							},
 						})
@@ -73,7 +72,8 @@ func PlanBuilder(from *tomledit.Document, to string) transform.Plan {
 							Block: kv.Block,
 							Name:  parser.Key{keys[0]},
 							Value: parser.MustValue(kv.Value),
-						})}
+						}),
+					}
 				} else if len(keys) > 1 {
 					step = transform.Step{
 						Desc: fmt.Sprintf("add %s key", kv.Key),
@@ -81,7 +81,8 @@ func PlanBuilder(from *tomledit.Document, to string) transform.Plan {
 							Block: kv.Block,
 							Name:  parser.Key{keys[1]},
 							Value: parser.MustValue(kv.Value),
-						})}
+						}),
+					}
 				}
 			default:
 				panic(fmt.Errorf("unknown diff type: %s", diff.Type))