+++ x/auth/tx/builder_test.go
@@ -5,6 +5,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	errorsmod "cosmossdk.io/errors"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/codec/legacy"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
@@ -165,7 +167,7 @@ func TestBuilderValidateBasic(t *testing.T) {
 	txBuilder.SetFeeAmount(badFeeAmount)
 	err = txBuilder.ValidateBasic()
 	require.Error(t, err)
-	_, code, _ := sdkerrors.ABCIInfo(err, false)
+	_, code, _ := errorsmod.ABCIInfo(err, false)
 	require.Equal(t, sdkerrors.ErrInsufficientFee.ABCICode(), code)
 
 	// require to fail validation when no signatures exist
@@ -174,7 +176,7 @@ func TestBuilderValidateBasic(t *testing.T) {
 	txBuilder.SetFeeAmount(feeAmount)
 	err = txBuilder.ValidateBasic()
 	require.Error(t, err)
-	_, code, _ = sdkerrors.ABCIInfo(err, false)
+	_, code, _ = errorsmod.ABCIInfo(err, false)
 	require.Equal(t, sdkerrors.ErrNoSignatures.ABCICode(), code)
 
 	// require to fail with nil values for tx, authinfo
@@ -189,7 +191,7 @@ func TestBuilderValidateBasic(t *testing.T) {
 
 	err = txBuilder.ValidateBasic()
 	require.Error(t, err)
-	_, code, _ = sdkerrors.ABCIInfo(err, false)
+	_, code, _ = errorsmod.ABCIInfo(err, false)
 	require.Equal(t, sdkerrors.ErrUnauthorized.ABCICode(), code)
 
 	require.Error(t, err)